build: Move images to be based on Amazon Linux 2023
We get a few advantages with moving to Amazon Linux 2023, the most
notable being that we get to use more packages provided by the OS rather
than downloading some. For example, we can use the node and npm binaries
from the RPM packages in AL2023.

We also take the opportunity to move to pulling base images from ECR
rather than DockerHub.

Using curl rather than wget also means we don't need to add wget to the
list of packages we're installing in the container, thus keeping things
nice and minimal.